# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Archive Tracking & Workflow Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /boxes:
    get:
      operationId: listBoxes
      tags: [boxes]
      summary: List all boxes
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of boxes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Box"
    post:
      operationId: createBox
      tags: [boxes]
      summary: Create a new box
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBoxBody"
      responses:
        "201":
          description: Created box
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Box"
  /boxes/{id}:
    get:
      operationId: getBox
      tags: [boxes]
      summary: Get a single box
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Box
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Box"
        "404":
          description: Not found
    put:
      operationId: updateBox
      tags: [boxes]
      summary: Update a box
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBoxBody"
      responses:
        "200":
          description: Updated box
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Box"
    delete:
      operationId: deleteBox
      tags: [boxes]
      summary: Delete a box
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /boxes/{id}/workflow:
    get:
      operationId: getBoxWorkflow
      tags: [workflow]
      summary: Get workflow steps for a box
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workflow steps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkflowStep"
    post:
      operationId: updateBoxWorkflow
      tags: [workflow]
      summary: Update workflow step status for a box
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkflowBody"
      responses:
        "200":
          description: Updated workflow step
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowStep"
        "403":
          description: Previous step not complete
  /boxes/{id}/ticket:
    get:
      operationId: getBoxTicket
      tags: [tickets]
      summary: Get e-ticket for a box
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ticket data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /scan/{ticketCode}:
    get:
      operationId: getByTicketCode
      tags: [tickets]
      summary: Get box info by ticket code (for QR scan)
      parameters:
        - name: ticketCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Box with workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResult"
        "404":
          description: Not found
  /roles:
    get:
      operationId: listRoles
      tags: [roles]
      summary: List all roles
      responses:
        "200":
          description: Roles list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
    post:
      operationId: createRole
      tags: [roles]
      summary: Create a role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoleBody"
      responses:
        "201":
          description: Created role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
  /roles/{id}:
    put:
      operationId: updateRole
      tags: [roles]
      summary: Update a role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoleBody"
      responses:
        "200":
          description: Updated role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
    delete:
      operationId: deleteRole
      tags: [roles]
      summary: Delete a role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Box:
      type: object
      properties:
        id:
          type: integer
        boxCode:
          type: string
        clientName:
          type: string
        collectionsOwner:
          type: string
          nullable: true
        placeOfOrigin:
          type: string
          nullable: true
        description:
          type: string
        location:
          type: string
        archiveYear:
          type: string
        totalItems:
          type: integer
        status:
          type: string
          enum: [received, in_progress, completed, returned]
        ticketCode:
          type: string
        notes:
          type: string
        inDate:
          type: string
          format: date-time
        outDate:
          type: string
          format: date-time
          nullable: true
        materialTypes:
          type: array
          items:
            type: string
            enum: [newspaper, maps, books, magazine, archives, heritage_items]
          nullable: true
        priority:
          type: string
          enum: [P0, P1, P2, P3]
          nullable: true
        custodyType:
          type: string
          enum: [loan, owned]
          nullable: true
        deadline:
          type: string
          format: date-time
          nullable: true
        cost:
          type: string
          nullable: true
        photoLink:
          type: string
          nullable: true
        currentStep:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - boxCode
        - clientName
        - status
        - ticketCode
        - currentStep
        - createdAt
        - updatedAt
    CreateBoxBody:
      type: object
      properties:
        boxCode:
          type: string
        clientName:
          type: string
        collectionsOwner:
          type: string
          nullable: true
        placeOfOrigin:
          type: string
          nullable: true
        description:
          type: string
        location:
          type: string
        archiveYear:
          type: string
        totalItems:
          type: integer
        notes:
          type: string
        photoLink:
          type: string
          nullable: true
        inDate:
          type: string
          format: date-time
        materialTypes:
          type: array
          items:
            type: string
            enum: [newspaper, maps, books, magazine, archives, heritage_items]
        priority:
          type: string
          enum: [P0, P1, P2, P3]
          nullable: true
        custodyType:
          type: string
          enum: [loan, owned]
          nullable: true
        deadline:
          type: string
          format: date-time
          nullable: true
        cost:
          type: number
          nullable: true
      required:
        - clientName
    UpdateBoxBody:
      type: object
      properties:
        boxCode:
          type: string
        clientName:
          type: string
        collectionsOwner:
          type: string
          nullable: true
        placeOfOrigin:
          type: string
          nullable: true
        description:
          type: string
        location:
          type: string
        archiveYear:
          type: string
        totalItems:
          type: integer
        notes:
          type: string
        status:
          type: string
          enum: [received, in_progress, completed, returned]
        outDate:
          type: string
          format: date-time
          nullable: true
        materialTypes:
          type: array
          items:
            type: string
            enum: [newspaper, maps, books, magazine, archives, heritage_items]
        priority:
          type: string
          enum: [P0, P1, P2, P3]
          nullable: true
        custodyType:
          type: string
          enum: [loan, owned]
          nullable: true
        deadline:
          type: string
          format: date-time
          nullable: true
        cost:
          type: number
          nullable: true
        photoLink:
          type: string
          nullable: true
    WorkflowStep:
      type: object
      properties:
        id:
          type: integer
        boxId:
          type: integer
        stepName:
          type: string
          enum: [cleaning, cataloging, scanning, qc, repacking, returning]
        stepOrder:
          type: integer
        status:
          type: string
          enum: [pending, in_progress, completed, skipped]
        assignedUserId:
          type: integer
          nullable: true
        assignedUserName:
          type: string
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - boxId
        - stepName
        - stepOrder
        - status
        - updatedAt
    UpdateWorkflowBody:
      type: object
      properties:
        stepName:
          type: string
          enum: [cleaning, cataloging, scanning, qc, repacking, returning]
        status:
          type: string
          enum: [pending, in_progress, completed, skipped]
        assignedUserId:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - stepName
        - status
    Ticket:
      type: object
      properties:
        boxId:
          type: integer
        boxCode:
          type: string
        clientName:
          type: string
        ticketCode:
          type: string
        qrCodeUrl:
          type: string
        description:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        archiveYear:
          type: string
          nullable: true
        totalItems:
          type: integer
          nullable: true
        inDate:
          type: string
          format: date-time
          nullable: true
        currentStep:
          type: string
        status:
          type: string
        workflowSteps:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowStep"
      required:
        - boxId
        - boxCode
        - clientName
        - ticketCode
        - qrCodeUrl
        - currentStep
        - status
        - workflowSteps
    ScanResult:
      type: object
      properties:
        box:
          $ref: "#/components/schemas/Box"
        workflowSteps:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowStep"
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
      required:
        - box
        - workflowSteps
        - users
    Role:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        workflowStep:
          type: string
          enum: [cleaning, cataloging, scanning, qc, repacking, returning, admin]
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - workflowStep
        - createdAt
    CreateRoleBody:
      type: object
      properties:
        name:
          type: string
        workflowStep:
          type: string
          enum: [cleaning, cataloging, scanning, qc, repacking, returning, admin]
        description:
          type: string
          nullable: true
      required:
        - name
        - workflowStep
    UpdateRoleBody:
      type: object
      properties:
        name:
          type: string
        workflowStep:
          type: string
          enum: [cleaning, cataloging, scanning, qc, repacking, returning, admin]
        description:
          type: string
          nullable: true
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          nullable: true
        roleId:
          type: integer
          nullable: true
        roleName:
          type: string
          nullable: true
        workflowStep:
          type: string
          nullable: true
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - active
        - createdAt
...