# lib/api-zod/src/generated/types/box.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Archive Tracking & Workflow Management API
 * OpenAPI spec version: 0.1.0
 */
import type { BoxCustodyType } from "./boxCustodyType";
import type { BoxMaterialTypesItem } from "./boxMaterialTypesItem";
import type { BoxPriority } from "./boxPriority";
import type { BoxStatus } from "./boxStatus";

export interface Box {
  id: number;
  boxCode: string;
  clientName: string;
  collectionsOwner?: string | null;
  placeOfOrigin?: string | null;
  description?: string;
  location?: string;
  archiveYear?: string;
  totalItems?: number;
  status: BoxStatus;
  ticketCode: string;
  notes?: string;
  inDate?: Date;
  outDate?: Date | null;
  materialTypes?: BoxMaterialTypesItem[] | null;
  priority?: BoxPriority;
  custodyType?: BoxCustodyType;
  deadline?: Date | null;
  cost?: string | null;
  photoLink?: string | null;
  currentStep: string;
  createdAt: Date;
  updatedAt: Date;
}