# lib/api-zod/src/generated/types/role.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Archive Tracking & Workflow Management API
 * OpenAPI spec version: 0.1.0
 */
import type { RoleWorkflowStep } from "./roleWorkflowStep";

export interface Role {
  id: number;
  name: string;
  workflowStep: RoleWorkflowStep;
  description?: string | null;
  createdAt: Date;
}