# lib/api-zod/src/generated/types/updateWorkflowBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Archive Tracking & Workflow Management API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateWorkflowBodyStatus } from "./updateWorkflowBodyStatus";
import type { UpdateWorkflowBodyStepName } from "./updateWorkflowBodyStepName";

export interface UpdateWorkflowBody {
  stepName: UpdateWorkflowBodyStepName;
  status: UpdateWorkflowBodyStatus;
  assignedUserId?: number | null;
  notes?: string | null;
}