# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Archive Tracking & Workflow Management API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  name: string;
  email?: string | null;
  roleId?: number | null;
  roleName?: string | null;
  workflowStep?: string | null;
  active: boolean;
  createdAt: Date;
}